+++ onnxruntime/core/mlas/inc/mlas.h
@@ -530,6 +530,8 @@ struct MLAS_GEMM_U8X8_SHAPE_PARAMS {
 };
 
 struct MLAS_GEMM_U8X8_DATA_PARAMS {
+    CBLAS_TRANSPOSE TransA = CblasNoTrans;
+    CBLAS_TRANSPOSE TransB = CblasNoTrans;
     const uint8_t* A = nullptr;
     size_t lda = 0;
     uint8_t ZeroPointA = 0;
