+++ onnxruntime/core/mlas/lib/qgemm_kernel_protocol.h
@@ -1,0 +1,573 @@
+/*++
+
+Copyright (c) Microsoft Corporation. All rights reserved.
+
+Licensed under the MIT License.
+
+Module Name:
+
+    qgemm_kernel_protocol.h
+
+Abstract:
+
+    This module defines the protocol to implement a kernel of 
+    quantized integer matrix/matrix multiply operation (QGEMM).
+
+--*/
+
+#pragma once
+
+#include "mlasi.h"
+
+template<typename KernelType>
+MLAS_FORCEINLINE
+bool
+MlasGemmU8X8TryGemvKernel(
+    const uint8_t* A,
+    const uint8_t* B,
+    size_t ldb,
+    int32_t* C,
+    size_t CountK,
+    size_t CountN,
+    bool BIsSigned
+)
+{
+    MLAS_UNREFERENCED_PARAMETER(A);
+    MLAS_UNREFERENCED_PARAMETER(B);
+    MLAS_UNREFERENCED_PARAMETER(ldb);
+    MLAS_UNREFERENCED_PARAMETER(C);
+    MLAS_UNREFERENCED_PARAMETER(CountK);
+    MLAS_UNREFERENCED_PARAMETER(CountN);
+    MLAS_UNREFERENCED_PARAMETER(BIsSigned);
+
+    return false;
+}
+
+template<typename KernelType>
+int32_t
+MlasGemmU8X8FixupZeroPointB(
+    int32_t ZeroPointB,
+    bool BIsSigned
+)
+{
+    MLAS_UNREFERENCED_PARAMETER(BIsSigned);
+
+    return ZeroPointB;
+}
+
+template<typename KernelType>
+MLAS_FORCEINLINE
+void
+MlasGemmU8X8FixupZeroPointB(
+    const uint8_t* PackedZeroPointB,
+    int32_t* ZeroPointBBuffer,
+    size_t N,
+    bool BIsSigned
+)
+{
+    int32_t ZeroPointB;
+
+    for (size_t n = 0; n < N; n++) {
+
+        ZeroPointB = typename KernelType::OffsetBType(PackedZeroPointB[n]);
+        ZeroPointB = MlasGemmU8X8FixupZeroPointB<KernelType>(ZeroPointB, BIsSigned);
+
+        ZeroPointBBuffer[n] = -ZeroPointB;
+    }
+
+    //
+    // Fill the misaligned slots of the zero point buffer with zeroes to guard
+    // against tools that check for uninitialized data usage.
+    //
+
+    size_t AlignedN = (N + MLAS_QGEMM_STRIDEN_THREAD_ALIGN - 1) & ~(MLAS_QGEMM_STRIDEN_THREAD_ALIGN - 1);
+
+    for (size_t n = N; n < AlignedN; n++) {
+        ZeroPointBBuffer[n] = 0;
+    }
+}
+
+template<typename KernelType>
+void
+MlasGemmU8X8CopyPackA(
+    typename KernelType::PackedAType* D,
+    const uint8_t* A,
+    size_t lda,
+    size_t CountM,
+    size_t CountK,
+    int32_t* RowSumBuffer
+);
+
+template<typename KernelType>
+void
+MlasGemmU8X8CopyPackB(
+    typename KernelType::PackedBType* D,
+    const uint8_t* B,
+    size_t ldb,
+    size_t CountN,
+    size_t CountK,
+    int32_t* ColumnSumBuffer,
+    bool BIsSigned
+);
+
+template<typename KernelType>
+size_t
+MlasGemmU8X8Kernel(
+    const typename KernelType::PackedAType* A,
+    const typename KernelType::PackedBType* B,
+    int32_t* C,
+    size_t PackedCountK,
+    size_t CountM,
+    size_t CountN,
+    size_t ldc,
+    const int32_t* RowSumBuffer,
+    const int32_t* ColumnSumBuffer,
+    const int32_t* ZeroPointB,
+    bool ZeroMode
+);
+
+
+template<typename KernelType>
+void
+MlasGemmU8X8Operation(
+    const MLAS_GEMM_U8X8_SHAPE_PARAMS* Shape,
+    const MLAS_GEMM_U8X8_DATA_PARAMS* Data,
+    const size_t RangeStartM,
+    const size_t RangeCountM,
+    const size_t RangeStartN,
+    const size_t RangeCountN
+    )
+/*++
+
+Routine Description:
+
+    This routine implements the quantized integer matrix/matrix multiply
+    operation (QGEMM).
+
+Arguments:
+
+    Shape - Supplies the structure containing the GEMM input and output shapes.
+
+    Data  - Supplies the structure containing the GEMM input and output data layout
+
+    RangeStartM - Supplies the starting row index to output.
+
+    RangeCountM - Supplies the number of rows to output.
+
+    RangeStartN - Supplies the starting column index to output.
+
+    RangeCountN - Supplies the number of columns to output.
+
+Return Value:
+
+    None.
+
+--*/
+{
+    constexpr MLAS_GEMM_U8X8_STRIDES Strides = KernelType::Strides;
+
+    MLAS_DECLSPEC_ALIGN(typename KernelType::PackedAType PanelA[Strides.M * Strides.K], 64);
+    MLAS_DECLSPEC_ALIGN(typename KernelType::PackedBType PanelB[Strides.N * Strides.K], 64);
+
+    MLAS_DECLSPEC_ALIGN(int32_t RowSumBuffer[Strides.M], 64);
+    MLAS_DECLSPEC_ALIGN(int32_t ColumnSumBuffer[Strides.N], 64);
+    MLAS_DECLSPEC_ALIGN(int32_t ZeroPointBBuffer[Strides.N], 64);
+
+    const size_t K = Shape->K;
+
+    const size_t lda = Data->lda;
+    const size_t ldb = Data->ldb;
+    const size_t ldc = Data->ldc;
+
+    const uint8_t* A = Data->A + RangeStartM * lda;
+    const uint8_t* B = (const uint8_t*)Data->B + RangeStartN;
+    int32_t* C = Data->C + RangeStartM * ldc + RangeStartN;
+    const uint8_t* PackedZeroPointB = Data->PerColumnZeroPoints ?
+        Data->ZeroPointB + RangeStartN : nullptr;
+
+    int32_t ZeroPointA = Data->ZeroPointA;
+    int32_t ZeroPointB = typename KernelType::OffsetBType(*Data->ZeroPointB);
+
+    //
+    // Try to use a GEMV kernel if supported by this kernel type.
+    //
+
+    if ((RangeCountM == 1) &&
+        (ZeroPointA == 0) && (PackedZeroPointB == nullptr) && (ZeroPointB == 0) &&
+        (Data->OutputProcessor == nullptr)) {
+        if (MlasGemmU8X8TryGemvKernel<KernelType>(A, B, ldb, C, K, RangeCountN, Shape->BIsSigned)) {
+            return;
+        }
+    }
+
+    //
+    // Fixup the sign bit of the per-matrix zero point offset of matrix B if the
+    // data is the opposite format of the kernel implementation. This value is
+    // ignored if per-column zero point offsets are used instead.
+    //
+
+    ZeroPointB = MlasGemmU8X8FixupZeroPointB<KernelType>(ZeroPointB, Shape->BIsSigned);
+
+    //
+    // Step through each slice of matrix B along the K dimension.
+    //
+
+    size_t CountK;
+
+    for (size_t k = 0; k < K; k += CountK) {
+
+        CountK = std::min(K - k, Strides.K);
+
+        const size_t PackedCountK = (CountK + KernelType::PackedK - 1) / KernelType::PackedK;
+
+        //
+        // Step through each slice of matrix B along the N dimension.
+        //
+
+        size_t CountN;
+
+        for (size_t n = 0; n < RangeCountN; n += CountN) {
+
+            CountN = std::min(RangeCountN - n, Strides.N);
+
+            //
+            // Fixup the sign bit of the per-column zero point offsets of matrix B
+            // if the data is the opposite format of the kernel implementation.
+            //
+
+            if (PackedZeroPointB != nullptr) {
+                MlasGemmU8X8FixupZeroPointB<KernelType>(
+                    PackedZeroPointB + n,
+                    ZeroPointBBuffer,
+                    CountN,
+                    Shape->BIsSigned);
+            }
+
+            //
+            // Copy a panel of matrix B to a local packed buffer.
+            //
+
+            MlasGemmU8X8CopyPackB<KernelType>(
+                PanelB,
+                B + n,
+                ldb,
+                CountN,
+                CountK,
+                ColumnSumBuffer,
+                Shape->BIsSigned);
+
+            MlasGemmU8X8ScaleSumBuffer(ColumnSumBuffer, CountN, -ZeroPointA);
+
+            //
+            // Step through each slice of matrix A along the M dimension.
+            //
+
+            int32_t* c = C + n;
+            size_t CountM;
+
+            for (size_t m = 0; m < RangeCountM; m += CountM) {
+
+                CountM = std::min(RangeCountM - m, Strides.M);
+
+                //
+                // Copy a panel of matrix A to a local packed buffer.
+                //
+
+                MlasGemmU8X8CopyPackA<KernelType>(
+                    PanelA,
+                    A + m * lda,
+                    lda,
+                    CountM,
+                    CountK,
+                    RowSumBuffer);
+
+                //
+                // Apply the global depth value constant without the ZeroPointB scaling from:
+                //
+                //     (A[i] - ZeroPointA) * (B[i] - ZeroPointB)
+                //              ==>
+                //     A[i] * B[i] - A[i] * ZeroPointB - B[i] * ZeroPointA + ZeroPointA * ZeroPointB
+                //
+                // The ZeroPointB term is factored out and either applied below for per-matrix
+                // quantization or inside the kernel for per-column quantization.
+                //
+
+                for (size_t mm = 0; mm < CountM; mm++) {
+                    RowSumBuffer[mm] -= int32_t(CountK) * ZeroPointA;
+                }
+
+                //
+                // Scale the row sums by the per-matrix zero point offset of matrix B.
+                //
+
+                if (PackedZeroPointB == nullptr) {
+                    MlasGemmU8X8ScaleSumBuffer(RowSumBuffer, CountM, -ZeroPointB);
+                }
+
+                //
+                // Step through the rows of the local packed buffer.
+                //
+
+                typename KernelType::PackedAType* pa = PanelA;
+                int32_t* RowSums = RowSumBuffer;
+                size_t RowsRemaining = CountM;
+
+                bool ZeroMode = (k == 0);
+                bool PostProcess = (k + CountK == K);
+
+                while (RowsRemaining > 0) {
+
+                    size_t RowsHandled = MlasGemmU8X8Kernel<KernelType>(
+                        pa,
+                        PanelB,
+                        c,
+                        PackedCountK,
+                        RowsRemaining,
+                        CountN,
+                        ldc,
+                        RowSums,
+                        ColumnSumBuffer,
+                        (PackedZeroPointB != nullptr) ? ZeroPointBBuffer : nullptr,
+                        ZeroMode);
+
+                    if (PostProcess && Data->OutputProcessor != nullptr) {
+                        Data->OutputProcessor->Process(
+                            Data->C,
+                            RangeStartM + m + CountM - RowsRemaining,
+                            RangeStartN + n,
+                            RowsHandled,
+                            CountN,
+                            Data->ldc);
+                    }
+
+                    c += ldc * RowsHandled;
+                    pa += KernelType::PackedK * PackedCountK * RowsHandled;
+                    RowSums += RowsHandled;
+                    RowsRemaining -= RowsHandled;
+                }
+            }
+        }
+
+        A += CountK;
+        B += CountK * ldb;
+    }
+}
+
+
+template<typename KernelType>
+void
+MlasGemmU8X8PackedOperation(
+    const MLAS_GEMM_U8X8_SHAPE_PARAMS* Shape,
+    const MLAS_GEMM_U8X8_DATA_PARAMS* Data,
+    const size_t RangeStartM,
+    const size_t RangeCountM,
+    const size_t RangeStartN,
+    const size_t RangeCountN
+    )
+/*++
+
+Routine Description:
+
+    This routine implements the quantized integer matrix/matrix multiply
+    operation (QGEMM).
+
+Arguments:
+
+    Shape - Supplies the structure containing the GEMM input and output shapes.
+
+    Data  - Supplies the structure containing the GEMM input and output data layout
+
+    RangeStartM - Supplies the starting row index to output.
+
+    RangeCountM - Supplies the number of rows to output.
+
+    RangeStartN - Supplies the starting column index to output.
+
+    RangeCountN - Supplies the number of columns to output.
+
+Return Value:
+
+    None.
+
+--*/
+{
+    constexpr MLAS_GEMM_U8X8_STRIDES Strides = KernelType::PackedStrides;
+
+    MLAS_DECLSPEC_ALIGN(typename KernelType::PackedAType PanelA[Strides.M * Strides.K], 64);
+
+    MLAS_DECLSPEC_ALIGN(int32_t RowSumBuffer[Strides.M], 64);
+    MLAS_DECLSPEC_ALIGN(int32_t ColumnSumBuffer[Strides.N], 64);
+    MLAS_DECLSPEC_ALIGN(int32_t ZeroPointBBuffer[Strides.N], 64);
+
+    const size_t K = Shape->K;
+
+    const size_t lda = Data->lda;
+    const size_t ldc = Data->ldc;
+
+    const uint8_t* A = Data->A + RangeStartM * lda;
+    const uint8_t* PackedB = (const uint8_t*)Data->B;
+    int32_t* C = Data->C + RangeStartM * ldc + RangeStartN;
+    const uint8_t* PackedZeroPointB = Data->PerColumnZeroPoints ?
+        Data->ZeroPointB + RangeStartN : nullptr;
+
+    int32_t ZeroPointA = Data->ZeroPointA;
+    int32_t ZeroPointB = typename KernelType::OffsetBType(*Data->ZeroPointB);
+
+    //
+    // Fixup the sign bit of the per-matrix zero point offset of matrix B if the
+    // data is the opposite format of the kernel implementation. This value is
+    // ignored if per-column zero point offsets are used instead.
+    //
+
+    ZeroPointB = MlasGemmU8X8FixupZeroPointB<KernelType>(ZeroPointB, Shape->BIsSigned);
+
+    //
+    // Extract the pointer to the column sum buffer from the packed matrix.
+    //
+
+    const size_t AlignedN =
+        (Shape->N + MLAS_QGEMM_STRIDEN_THREAD_ALIGN - 1) & ~(MLAS_QGEMM_STRIDEN_THREAD_ALIGN - 1);
+    const int32_t* PackedColumnSumBuffer = (const int32_t*)PackedB;
+    PackedB = (const uint8_t*)(PackedColumnSumBuffer + AlignedN);
+    PackedColumnSumBuffer += RangeStartN;
+
+    //
+    // Step through each slice of matrix B along the K dimension.
+    //
+
+    size_t CountK;
+
+    for (size_t k = 0; k < K; k += CountK) {
+
+        CountK = std::min(K - k, Strides.K);
+
+        const size_t PackedCountK = (CountK + KernelType::PackedK - 1) / KernelType::PackedK;
+
+        if (k > 0) {
+            std::fill_n(ColumnSumBuffer, Strides.N, 0);
+        }
+
+        //
+        // Step through each slice of matrix B along the N dimension.
+        //
+
+        size_t CountN;
+
+        for (size_t n = 0; n < RangeCountN; n += CountN) {
+
+            CountN = std::min(RangeCountN - n, Strides.N);
+
+            if (k == 0) {
+                MlasGemmU8X8ScaleSumBuffer(ColumnSumBuffer, PackedColumnSumBuffer + n,
+                    CountN, -ZeroPointA);
+            }
+
+            //
+            // Fixup the sign bit of the per-column zero point offsets of matrix B
+            // if the data is the opposite format of the kernel implementation.
+            //
+
+            if (PackedZeroPointB != nullptr) {
+                MlasGemmU8X8FixupZeroPointB<KernelType>(
+                    PackedZeroPointB + n,
+                    ZeroPointBBuffer,
+                    CountN,
+                    Shape->BIsSigned);
+            }
+
+            //
+            // Step through each slice of matrix A along the M dimension.
+            //
+
+            const uint8_t* b = PackedB + (RangeStartN + n) *
+                KernelType::PackedK * PackedCountK;
+            int32_t* c = C + n;
+            size_t CountM;
+
+            for (size_t m = 0; m < RangeCountM; m += CountM) {
+
+                CountM = std::min(RangeCountM - m, Strides.M);
+
+                //
+                // Copy a panel of matrix A to a local packed buffer.
+                //
+
+                MlasGemmU8X8CopyPackA<KernelType>(
+                    PanelA,
+                    A + m * lda,
+                    lda,
+                    CountM,
+                    CountK,
+                    RowSumBuffer);
+
+                //
+                // Apply the global depth value constant without the ZeroPointB scaling from:
+                //
+                //     (A[i] - ZeroPointA) * (B[i] - ZeroPointB)
+                //              ==>
+                //     A[i] * B[i] - A[i] * ZeroPointB - B[i] * ZeroPointA + ZeroPointA * ZeroPointB
+                //
+                // The ZeroPointB term is factored out and either applied below for per-matrix
+                // quantization or inside the kernel for per-column quantization.
+                //
+
+                for (size_t mm = 0; mm < CountM; mm++) {
+                    RowSumBuffer[mm] -= int32_t(CountK) * ZeroPointA;
+                }
+
+                //
+                // Scale the row sums by the per-matrix zero point offset of matrix B.
+                //
+
+                if (PackedZeroPointB == nullptr) {
+                    MlasGemmU8X8ScaleSumBuffer(RowSumBuffer, CountM, -ZeroPointB);
+                }
+
+                //
+                // Step through the rows of the local packed buffer.
+                //
+
+                typename KernelType::PackedAType* pa = PanelA;
+                int32_t* RowSums = RowSumBuffer;
+                size_t RowsRemaining = CountM;
+
+                bool ZeroMode = (k == 0);
+                bool PostProcess = (k + CountK == K);
+
+                while (RowsRemaining > 0) {
+
+                    size_t RowsHandled = MlasGemmU8X8Kernel<KernelType>(
+                        pa,
+                        b,
+                        c,
+                        PackedCountK,
+                        RowsRemaining,
+                        CountN,
+                        ldc,
+                        RowSums,
+                        ColumnSumBuffer,
+                        (PackedZeroPointB != nullptr) ? ZeroPointBBuffer : nullptr,
+                        ZeroMode);
+
+                    if (PostProcess && Data->OutputProcessor != nullptr) {
+                        Data->OutputProcessor->Process(
+                            Data->C,
+                            RangeStartM + m + CountM - RowsRemaining,
+                            RangeStartN + n,
+                            RowsHandled,
+                            CountN,
+                            Data->ldc);
+                    }
+
+                    c += ldc * RowsHandled;
+                    pa += KernelType::PackedK * PackedCountK * RowsHandled;
+                    RowSums += RowsHandled;
+                    RowsRemaining -= RowsHandled;
+                }
+            }
+        }
+
+        A += CountK;
+        PackedB = (const uint8_t*)PackedB + AlignedN * CountK;
+    }
+}
